refactor: 주문 응용 command DTO 패키지 수정
